Fix autocomplete select bug hidding the form
Events were binded to hide the booking bottom row on clicking anywhere inside the document. To prevent clicks inside the widget to trigger this, a stopPropagation is used.

The autocomplete widget stopped working because of this, but adding stopPropagation to the select event fixed it. Don't know how or why.